Remove outdated stats

+++ Code/Experiments/exp_stats.h
@@ -4,7 +4,7 @@
 #define EXP_STATS
 
 extern double best_score;
-extern int correct_count;
+extern double correct_count;
 extern double question_num;
 extern double return_size;
